DaliugeApplications reusage - first iteration
At the moment we are storing each application drop's state using an
app_data structure, which is created in each DaliugeApplication.init()
method, then set into the dlg_app_info structure managed by the python
code, and then read from the DaliugeApplication.run() method, instead of
using the DaliugeApplication instance itself to store its state. This set
of changes changes how this works, removing the need to re-instantiate
an DaliugeApplication, and saving *it* instead as the dlg_app_info.data
member.

On top of this, the dlg_app_info is also now more invisible to children
classes, with some utility protected methods provided by the base class
to deal with inputs, outputs, etc.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>