Unload unversioned askapsoft modules
Builds started to fail during the weekend because we were unload a
particular version of the askapsoft modules which does not exist anymore
(has been upgraded from 0.20.3 to 0.21.0), and therefore we were running
into old problems. Unloading the unversioned flavour of the module will
work more reliably and save us from future accidents like this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>