New example script on how to write with ADIOS2
This script shows us how to write columns with casacore-enabled ADIOS2 using
the Adios2StMan storage manager. This runs well both in a single-rank MPI
communicator, or in multiple-rank communicators. I have only tried it in a
single node, but there's no reason why it shouldn't work across many nodes.

The script allows users to specify the column type (only a reduced set of
datatypes for the time being), and whether to use all MPI ranks or a subset of
them (the even ones) to perform the writing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>