Close MS after final writing
When running as a standalone program all objects are correctly
dereferenced and their destructors are invoked, including the
OSKAR measurement set object, which at destruction time effectively
clsoes the measurement set. In the context of a dlg execution this is
not the case though, as drops are not immediately deleted until a
session is deleted.

Let's therefore be good citizens and let's explicitly close our
underlying measurement set as soon as we are done writing into it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>