Using an explicit MPI communicator
The object stored under dlg.mpi_comm is an MPI communicator created by
the start_dfms_cluster when the -M option is specified. This
communicator contains only the ranks hosting node managers, and
therefore is what we want to use when issuing MPI collective calls
through ADIOS/casacore, which would otherwise fail due to some ranks in
the WORLD communicator not participating.

After taking this communicator from dlg.mpi_comm, we pass it down to the
OSKAR MS creation routine, which in turn will pass it both to the
Adios2StMan instance and to all Table object creation routines that
happen as part of the MS creation.

With this, our ingest pipeline example is (finally!) successfully
producing an ADIOS2-stored measurement set.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>