AveragerSinkDrop improvements
The number of streams that need to be setup by the receiver depends on
the number of drop's inputs, which we now at runtime, and that therefore
doesn't need to be given as an input at drop creation time.

I'm also exposing the option to use adios or not as a drop parameter for
easy toggling.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>