Porting pipeline scripts to new remote mechanisms
Daliuge now offers the option to choose the mechanism to use for remote
coordination between nodes (MPI-based or SLURM-based). In the case of
the ingest pipeline we can use either, so it's up to the user to choose
(we default to MPI, like it was before). In the case of the imaging
pipeline we cannot use MPI though, so users are not given the option and
SLURM is based (in SLURM-based systems).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>