Store baseline exclusion indices in set, not list
This object is used later for filtering baselines based on their
indices, which requires a lookup operation. Lookups on lists are known
to be slow due to their O(N) nature, so a set() fits the bill much
better in this situation.

This seems to have been the cause for the big slowdown we started
experiencing when running experiments with the AA4 telescope model,
where the slowness of the system became apparent.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>