Adjusting cimager drop to latest dlg changes
With the latest metadata attribute changes in daliuge there are a number
of things that have changed, in particular on the BashApp. On the one
hand, the private attribute holding the command to be run is now called
"command" instead of "_command", so that needed adjusting. More
importantly, children classes now cannot use the kwargs argument during
initialize() to pass down values for "command", as the self.command
attribute in the BashApp (and all daliuge attributes in general) is
initialized at __init__() time before initialize() is called.

I also took the opportunity to pass the correct type to the super() call
in CImagerDrop's initialize method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>