Requesting reasonable number of CPUs
Until now with SLURM we have not been explicit on the amount of CPUs we
need. In most systems this means that we'll get one CPU per task
(therefore by node), and thus we have been running, in the case of NMs
hosting the averger sink, N OSKAR instances, plus two averaging steps,
using a single CPU. Thus it shouldn't be a surprise we have seen bad
performances in some systems.

We definitely want to make sure we are doing things correctly in SUMMIT
too, but that's up to Jason to find out.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>