First try at using job arrays for imaging
A new mode of operation is being now introduced for the imaging
pipeline. Instead of using daliuge, logical graphs, etc, we are
resorting to use job arrays to make everyone's life easier, and to be
actually able to produce images in SUMMIT.

The new mode is the default, but the old mode (daliuge-based) can be
reached via the -D command line switch. This code has been tested in
SLURM, and i *think* it should work in LFS as well, but that is for
Jason to assess.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>