Making new processes more fork()-friendly
New processes started via multiprocessing.Process() are fork()-ed off
the parent, meaning that they inherit all its memory and state. This
includes lock states, which can lead to deadlocks when trying to acquire
them in the children processes. A nice explanation of this can be found
in https://codewithoutrules.com/2018/09/04/python-multiprocessing/.

This situation is indeed what seems to be happening within the logging
module, as even the simples of log outputs cannot be produced by the
children processes *in some situations*, although very rarely. To avoid
this particular possible threat (which seems real, given the experiments
I've conducted in our Nimbix infrastructure) the first thing I'm doing
now in the children processes is removing all the existing handlers, as
they might be in a locked state, and reconfiguring the logging system
altogether to look as similar as it was before. We are currently loosing
the ability to put the logs coming from the children processes into the
individual dlgNM.log files, but we can live with that -- and we can work
that out later actually, if required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>