Revised default frequency-related inputs for ingest pipeline
Now that we are starting to run the pipeline in more nodes we found that
our frequency coverage by default is not great. We sat down with Richard
and worked out how to cover the frequency ranges of both the
"interesting" EOR data, and the "final galaxy" on the foreground in a
full SUMMIT run. In a full pipeline we will have 27642 GPUs to simulate
data, so using a step size of 6.4 kHz and a starting frequency of 45.9912
MHz we can cover the frequency range up to 222.9 MHz.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

lala