OSKAR instances started with Popen
This ensures that we use execvp underneath instead of fork(), which is
the cause of our sufferings. We loose on the other hand the ability to
easily transport python objects between the two processes, but that's a
small price to pay.

Together with this change I'm also modifying where files needed by the
OSKAR simulators are written. Before we put them into the temporary
filesystem area, but Markus has asked a few times to preserve them, and
thus I'm hacking my way to deduce a oskar_log_dir location (defaults to
'.') where all OSKAR-related outputs will reside. I'm putting there the
.ini file, the spead configuration dumps, and also letting the OSKAR
instance log into a file in that directory.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>