Make time comparissons global, not local
Time values on which the heatmap is based were actually only valid
within each row, as their values were calculated relative to the start
time of the corresponding node. This is misleading if we are showing all
the information together, as a single plot implies that the time axis is
the same for all.

After trying out for a couple of output directories I could see little
difference though, indicating that all node managers are starting the
execution of their parts of the graph more-or-less at the same time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>