First adjustments to multi-process job arrays
This commit simplifies the logic of the runtime script of the
job-array-based imaging pipeline. It removes an extra increment on one
of the input values, renames the variable 'r' to 'rank' to give it more
significance (although it's not 100% accurate), and also adds a final
wait so the script doesn't finish before the actual cimager instances
are done.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>