Surround OSKAR execution with try/except
This way we can log any exceptions we experience, whatever they might
be. These logged exceptions will then appear in the logfiles we generate
when configuring the logging system to produce logfiles (e.g., in the
case of actual runs).

There is also no need to call finalize(), as this is called internally
by run() already.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>