Abort mechanism for OSKAR simulations
OSKAR's python bindings use threads to execution the actual simulation
steps, and yet a separate thread to "finalize" a block of simulated
visibilities, which is where we send them through SPEAD. An error on
this last thread then doesn't propagate to the simulation threads, who
effectively hang forever waiting for the "finaliser" thread to send them
a signal indicating it's done with its work.

This means we need to take care of exceptions ourselves. By catching
exceptions in process_block we now flag whether we need to "abort",
which in our case means doing nothing for the rest of the invocations
from the Interferometer class. After running the Interferometer class we
also check the abort flag, and exit with code 2 to signal that we
aborted prematurely.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>