Adding support for "dlg" remote in SUMMIT
This new remote mechanism allows us to fire up the individual managers
of the daliuge cluster without them being bound to a single MPI context,
while still being able to communicate with each other. This should
prevent whole jobs from failing even in the presence of segmentation
faults in one of the nodes.

We have tested this with 6 and 200 nodes, and it has successfully run to
completion.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>