Rows sorted by time, rank, and baseline
Now that we are aiming to write visibilities in parallel we need to
organize the different ranks so they write in parallel correctly. We do
this by writing all baselines in a rank as a single block, and then
having all ranks write them consecutively in rank order, for a given
time step.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>